Removed the double div situation
